requirements.txt
Flask==2.3.3
mysql-connector-python==8.1.0
requests==2.31.0
bcrypt==4.0.1
python-dotenv==1.0.0 